+++ cli/cli.rb
@@ -3,10 +3,11 @@ require 'thor'
 
 $LOAD_PATH << File.dirname(__FILE__)
 
+require 'dinghy/check_env'
+require 'dinghy/dnsmasq'
+require 'dinghy/http_proxy'
 require 'dinghy/unfs'
 require 'dinghy/vagrant'
-require 'dinghy/dnsmasq'
-require 'dinghy/check_env'
 
 class DinghyCLI < Thor
   option :memory,
@@ -20,6 +21,10 @@ class DinghyCLI < Thor
   option :provider,
     aliases: :p,
     desc: "which Vagrant provider to use, only takes effect when initializing a new VM"
+  option :proxy,
+    type: :boolean,
+    default: false,
+    desc: "start the HTTP proxy as well"
   desc "up", "start the Docker VM and services"
   def up
     vagrant = Vagrant.new
@@ -30,6 +35,9 @@ class DinghyCLI < Thor
     vagrant.install_docker_keys
     Dnsmasq.new.up
     CheckEnv.new.run
+    if options[:proxy]
+      HttpProxy.new.up
+    end
   end
 
   desc "ssh [args...]", "run vagrant ssh on the VM"
@@ -39,9 +47,10 @@ class DinghyCLI < Thor
 
   desc "status", "get VM and services status"
   def status
-    puts " VM: #{Vagrant.new.status}"
-    puts "NFS: #{Unfs.new.status}"
-    puts "DNS: #{Dnsmasq.new.status}"
+    puts "  VM: #{Vagrant.new.status}"
+    puts " NFS: #{Unfs.new.status}"
+    puts " DNS: #{Dnsmasq.new.status}"
+    puts "HTTP: #{HttpProxy.new.status}"
   end
 
   desc "halt", "stop the VM and services"